requirements.txt
hydra-core
torch 
numpy
pyrr 
trimesh